refactor(api): migrate organization-tags.repository.js
