+++ api/scripts/create-organizations-with-tags-and-target-profiles.js
@@ -8,9 +8,9 @@ import { disconnect } from '../db/knex-database-connection.js';
 import { createOrganizationsWithTagsAndTargetProfiles } from '../lib/domain/usecases/create-organizations-with-tags-and-target-profiles.js';
 import * as organizationValidator from '../lib/domain/validators/organization-with-tags-and-target-profiles-script.js';
 import { DomainTransaction as domainTransaction } from '../lib/infrastructure/DomainTransaction.js';
-import * as organizationTagRepository from '../lib/infrastructure/repositories/organization-tag-repository.js';
 import * as targetProfileShareRepository from '../lib/infrastructure/repositories/target-profile-share-repository.js';
 import * as dataProtectionOfficerRepository from '../src/organizational-entities/infrastructure/repositories/data-protection-officer.repository.js';
+import * as organizationTagRepository from '../src/organizational-entities/infrastructure/repositories/organization-tag.repository.js';
 import { tagRepository } from '../src/organizational-entities/infrastructure/repositories/tag.repository.js';
 import * as organizationRepository from '../src/shared/infrastructure/repositories/organization-repository.js';
 import { temporaryStorage } from '../src/shared/infrastructure/temporary-storage/index.js';
