+++ api/tests/organizational-entities/integration/domain/usecases/get-recently-used-tags.usecase.test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import * as organizationTagRepository from '../../../../../lib/infrastructure/repositories/organization-tag-repository.js';
 import { Tag } from '../../../../../src/organizational-entities/domain/models/Tag.js';
 import { getRecentlyUsedTags } from '../../../../../src/organizational-entities/domain/usecases/get-recently-used-tags.usecase.js';
+import * as organizationTagRepository from '../../../../../src/organizational-entities/infrastructure/repositories/organization-tag.repository.js';
 import { databaseBuilder, expect } from '../../../../test-helper.js';
 
 describe('Integration | Organizational Entities | Domain | UseCase | get-recently-used-tags', function () {
